ShowViewController 메소드 추가
* ViewController 보여주는 코드 메소드로 수정
