+++ Modakyi/Method/NetworkCheck.swift
