+++ Modakyi/Unused/UnusedViewController.swift
@@ -15,7 +15,7 @@ class UnusedViewController: UIViewController {
     var unusedTexts = [Int]()
     
     let uid = Auth.auth().currentUser?.uid
-
+    
     @IBOutlet weak var collectionview: UICollectionView!
     @IBOutlet weak var labelView: UIView!
     @IBOutlet weak var indicatorView: UIActivityIndicatorView!
@@ -33,7 +33,7 @@ class UnusedViewController: UIViewController {
                 let textData = try JSONDecoder().decode([String: StudyStimulateText].self, from: jsonData)
                 let texts = Array(textData.values)
                 self.studyStimulateTexts = texts.sorted { Int($0.id)! > Int($1.id)! }
-//                print("Unused - studyStimulateText: \(self.studyStimulateTexts)")
+                //                print("Unused - studyStimulateText: \(self.studyStimulateTexts)")
             } catch let error {
                 print("ERROR JSON Parsing \(error.localizedDescription)")
             }
@@ -44,7 +44,7 @@ class UnusedViewController: UIViewController {
             guard let value = snapshot.value as? [Int] else {
                 // 사용한 글귀가 없으니까 전체 글귀 보여주기
                 self.unusedTexts = self.studyStimulateTexts.map { Int($0.id)! }
-//                print("Unused 미사용 글귀 id: \(self.unusedTexts)")
+                //                print("Unused 미사용 글귀 id: \(self.unusedTexts)")
                 
                 DispatchQueue.main.async {
                     self.collectionview.reloadData()
@@ -115,10 +115,7 @@ extension UnusedViewController: UICollectionViewDataSource {
 extension UnusedViewController: UICollectionViewDelegate {
     // 셀 클릭했을 때
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        let storyboard = UIStoryboard(name: "Main", bundle: Bundle.main)
-        guard let detailViewController = storyboard.instantiateViewController(withIdentifier: "DetailViewController") as? DetailViewController else { return }
-        detailViewController.id = String(unusedTexts[indexPath.row])
-        self.present(detailViewController, animated: true, completion: nil)
+        presentDetailViewController(self, String(unusedTexts[indexPath.row]))
     }
 }
